+++ packages/adapter-cloudflare-workers/index.d.ts
@@ -1,3 +1,3 @@
 import { Adapter } from '@sveltejs/kit';
 
-export default function plugin(): Adapter;
+export default function plugin(options: { config?: string }): Adapter;