+++ packages/kit/src/runtime/server/endpoint.js
@@ -54,7 +54,6 @@ export async function render_endpoint(event, mod, state) {
 		}
 
 		if (state.prerendering) {
-			response.headers.set('x-sveltekit-routeid', /** @type {string} */ (event.routeId));
 			response.headers.set('x-sveltekit-prerender', String(prerender));
 		}