+++ packages/kit/test/apps/basics/src/routes/load/raw-body/+page.svelte
@@ -1,7 +1,3 @@
-<script>
-	/** @type {import('./$types').PageData} */
-	export let data;
-</script>
-
-<pre class="parsed">{JSON.stringify(data.body)}</pre>
-<pre class="raw">{data.rawBody}</pre>
+<a href="/load/raw-body/dataview">DataView</a>
+<a href="/load/raw-body/string">String</a>
+<a href="/load/raw-body/uint8array">Uint8Array</a>