+++ packages/kit/test/apps/basics/src/routes/load/serialization.json/+server.js
@@ -1,8 +1,0 @@
-import { json } from '@sveltejs/kit';
-
-/** @type {import('./$types').RequestHandler} */
-export function GET() {
-	return json({
-		answer: 42
-	});
-}