+++ packages/migrate/migrations/routes/utils.js
@@ -504,11 +504,11 @@ export function rewrite_type(node, code, old_type, new_type) {
  * @param {NonNullable<ReturnType<typeof parse>>} file
  */
 export function uppercase_migration(methods, file) {
-	const old_methods = ['get', 'post', 'put', 'patch', 'del'].filter((name) =>
-		file.exports.map.has(name)
+	const old_methods = new Set(
+		['get', 'post', 'put', 'patch', 'del'].filter((name) => file.exports.map.has(name))
 	);
 
-	if (old_methods.length && !methods.length) {
+	if (old_methods.size && !methods.length) {
 		for (const statement of file.ast.statements) {
 			for (const method of old_methods) {
 				const fn = get_function_node(
@@ -516,16 +516,17 @@ export function uppercase_migration(methods, file) {
 					/** @type{string} */ (file.exports.map.get(method))
 				);
 				if (!fn?.name) {
-					return;
+					continue;
 				}
 				file.code.overwrite(
 					fn.name.getStart(),
 					fn.name.getEnd(),
 					method === 'del' ? 'DELETE' : method.toUpperCase()
 				);
+				old_methods.delete(method);
 			}
 		}
 	}
 
-	return file.code.toString();
+	return old_methods.size ? undefined : file.code.toString();
 }