+++ packages/adapter-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare-workers
 
+## 1.0.0-next.59
+
+### Patch Changes
+
+- Use config.kit.paths.base prefix for static assets ([#4448](https://github.com/sveltejs/kit/pull/4448))
+
+## 1.0.0-next.58
+
+### Patch Changes
+
+- Add config option to set custom `wrangler.toml` file name ([#7104](https://github.com/sveltejs/kit/pull/7104))
+
 ## 1.0.0-next.57
 
 ### Patch Changes