+++ packages/adapter-netlify/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sveltejs/adapter-netlify
 
+## 1.0.0-next.81
+
+### Patch Changes
+
+- Use config.kit.paths.base prefix for static assets ([#4448](https://github.com/sveltejs/kit/pull/4448))
+
+## 1.0.0-next.80
+
+### Patch Changes
+
+- Transfer server data as devalue-encoded JSON ([#7177](https://github.com/sveltejs/kit/pull/7177))
+
+## 1.0.0-next.79
+
+### Patch Changes
+
+- [fix] clear out build directory when running `adapter-netlify` ([#7159](https://github.com/sveltejs/kit/pull/7159))
+
 ## 1.0.0-next.78
 
 ### Patch Changes
@@ -64,7 +82,7 @@
 
 ## 1.0.0-next.68
 
-### Minor Changes
+### Patch Changes
 
 - feat: get client IP from Context object rather than the request header ([#5473](https://github.com/sveltejs/kit/pull/5473))